Add support for file multiupload and deletion of FileReferences from an ObjectStorage property.
This patch adds two new ViewHelpers:
h:form.multiupload can be used to render a multiupload form and a list of all file references in a property.
h:form.multiupload.delete renders a checkbox that, when unchecked, removes the file reference from the object storage.

Notice: This is only a draft and not considered stable. There is still one mayor bug, a few TODOs and some untested use cases...